1. 버튼 중복 클릭 방지 코드 추가 2. 어플 색상 통일감있게 변경 3. 달력에서 TEST 카운트 얻어오는 쿼리 수정

+++ forms_plus/forms_plus/forms_plus/RankingStage1Page.xaml.cs
@@ -13,6 +13,7 @@ namespace forms_plus
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class RankingStage1Page : ContentPage
     {
+        private bool accessible = true;
         public RankingStage1Page()
         {
             InitializeComponent();
@@ -28,12 +29,25 @@ namespace forms_plus
 
         private async void Btn_DeleteHistory(object sender, EventArgs e)
         {
-            bool anwser = await DisplayAlert("기록 지우기", "기록을 모두 지우시겠습니까?", "네", "아니오");
-
-            if (anwser == true)
+            try
             {
-                await App.RkInfoDatabase.DeleteStageAsync("1");
-                listx.ItemsSource = await App.RkInfoDatabase.GetRankingsAsync("1");
+                if (accessible == true)
+                {
+                    accessible = false;
+                    bool anwser = await DisplayAlert("기록 지우기", "기록을 모두 지우시겠습니까?", "네", "아니오");
+
+                    if (anwser == true)
+                    {
+                        await App.RkInfoDatabase.DeleteStageAsync("1");
+                        listx.ItemsSource = await App.RkInfoDatabase.GetRankingsAsync("1");
+                    }
+                    accessible = true;
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                accessible = true;
             }
         }        
     }
