+++ forms_plus/forms_plus/forms_plus/RankingStage3Page.xaml.cs
@@ -12,6 +12,7 @@ namespace forms_plus
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class RankingStage3Page : ContentPage
     {
+        private bool accessible = true;
         public RankingStage3Page()
         {
             InitializeComponent(); 
@@ -26,12 +27,25 @@ namespace forms_plus
 
         private async void Btn_DeleteHistory(object sender, EventArgs e)
         {
-            bool anwser = await DisplayAlert("기록 지우기", "기록을 모두 지우시겠습니까?", "네", "아니오");
-
-            if (anwser == true)
+            try
             {
-                await App.RkInfoDatabase.DeleteStageAsync("3");
-                listx.ItemsSource = await App.RkInfoDatabase.GetRankingsAsync("3");
+                if (accessible == true)
+                {
+                    accessible = false;
+                    bool anwser = await DisplayAlert("기록 지우기", "기록을 모두 지우시겠습니까?", "네", "아니오");
+
+                    if (anwser == true)
+                    {
+                        await App.RkInfoDatabase.DeleteStageAsync("3");
+                        listx.ItemsSource = await App.RkInfoDatabase.GetRankingsAsync("3");
+                    }
+                    accessible = true;
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                accessible = true;
             }
         }
     }
